changed naming in base-classes and repalaced float with double

+++ nikonov.andrew/T1/additional-utilities.hpp
@@ -6,7 +6,7 @@ namespace nikonov
 {
   void fillShapeCollection(Shape* collection[], size_t& cnt, size_t& noncorrect);
   void destoy(Shape* collection[], size_t& cnt);
-  void ispScale(Shape* shp, float x, float y, float k);
+  void ispScale(Shape* shp, double  x, double  y, double  k);
   bool processCollection(Shape* collection[], size_t cnt);
 }
 #endif
